! need to be fixed LIE -> UserMapper & others

+++ src/main/java/ru/xpendence/streamcast/dto/AbstractDto.java
@@ -41,5 +41,5 @@ public abstract class AbstractDto implements Serializable {
     private LocalDateTime created;
 
     @JsonView(value = {View.DoNotShow.class})
-    private Long active;
+    private Integer active;
 }
